refactor(CarViewTest,CarGameTest): 유효성 검사 테스트를 CarGameTest에서 CarViewTest로 이동
